version 20:1.add start btn when challenge,2.slide stop,cancel slide rebound

+++ Client/MineCC/assets/Scripts/Socket.js
@@ -34,7 +34,7 @@ var creatWS = function (argument) {
         var cmd = data.substring(0, i1);
         var sResponse = data.substring(i1+1);
         if (cmd == GLB.GETVERSION){
-            var iVersion = parseInt(sResponse);
+            var iVersion = parseFloat(sResponse);
             console.log("iVersion = ", iVersion);
             if (iVersion - GLB.iVersion > 0){
                 GLB.bHotUpdate = true;
